Move Response into SearchResponderV2
