--- webapi/src/main/scala/org/knora/webapi/messages/v2/responder/searchmessages/SearchMessagesV2.scala
@@ -1,43 +1,0 @@
-/*
- * Copyright © 2021 - 2023 Swiss National Data and Service Center for the Humanities and/or DaSCH Service Platform contributors.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package org.knora.webapi.messages.v2.responder.searchmessages
-
-import org.knora.webapi.ApiV2Schema
-import org.knora.webapi.SchemaOption
-import org.knora.webapi.config.AppConfig
-import org.knora.webapi.core.RelayedMessage
-import org.knora.webapi.messages.OntologyConstants
-import org.knora.webapi.messages.ResponderRequest.KnoraRequestV2
-import org.knora.webapi.messages.util.rdf.JsonLDDocument
-import org.knora.webapi.messages.util.rdf.JsonLDInt
-import org.knora.webapi.messages.util.rdf.JsonLDObject
-import org.knora.webapi.messages.util.rdf.JsonLDString
-import org.knora.webapi.messages.v2.responder.*
-
-sealed trait SearchResponderRequestV2 extends KnoraRequestV2 with RelayedMessage
-
-/**
- * Represents the number of resources found by a search query.
- */
-case class ResourceCountV2(numberOfResources: Int) extends KnoraJsonLDResponseV2 {
-  override def toJsonLDDocument(
-    targetSchema: ApiV2Schema,
-    appConfig: AppConfig,
-    schemaOptions: Set[SchemaOption]
-  ): JsonLDDocument =
-    JsonLDDocument(
-      body = JsonLDObject(
-        Map(
-          OntologyConstants.SchemaOrg.NumberOfItems -> JsonLDInt(numberOfResources)
-        )
-      ),
-      context = JsonLDObject(
-        Map(
-          "schema" -> JsonLDString(OntologyConstants.SchemaOrg.SchemaOrgPrefixExpansion)
-        )
-      )
-    )
-}
