+++ webapi/src/main/scala/org/knora/webapi/routing/v2/SearchRouteV2.scala
@@ -22,7 +22,7 @@ import org.knora.webapi.messages.OntologyConstants
 import org.knora.webapi.messages.SmartIri
 import org.knora.webapi.messages.ValuesValidator
 import org.knora.webapi.messages.util.search.gravsearch.GravsearchParser
-import org.knora.webapi.messages.v2.responder.searchmessages.*
+import org.knora.webapi.responders.v2.ResourceCountV2
 import org.knora.webapi.responders.v2.SearchResponderV2
 import org.knora.webapi.routing.Authenticator
 import org.knora.webapi.routing.RouteUtilV2
